requirements.txt
Django==4.2.7
Pillow==10.0.1
django-ckeditor==6.7.0
django-taggit==4.0.0
gunicorn==21.2.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
